# user-sample-api/swagger.json
{"openapi":"3.0.0","info":{"title":"User API","description":"Open API documentation for the User API","version":"1.0","contact":{}},"tags":[],"servers":[],"components":{"securitySchemes":{"bearer":{"scheme":"bearer","bearerFormat":"JWT","type":"http"}},"schemas":{"User":{"type":"object","properties":{"id":{"type":"string"},"firstName":{"type":"string"},"lastName":{"type":"string"},"address":{"type":"string"},"email":{"type":"string"},"password":{"type":"string"}},"required":["id","firstName","lastName","address","email","password"]}}},"paths":{"/users":{"post":{"operationId":"UserController_createUser","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/User"}}}},"responses":{"201":{"description":"Created user succesfully","content":{"application/json":{"schema":{"$ref":"#/components/schemas/User"}}}},"400":{"description":"Something went wrong with the user creation"}}},"get":{"operationId":"UserController_getAllUsers","parameters":[],"responses":{"200":{"description":"User object","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/User"}}}}},"400":{"description":"Something went wrong with the users retrieval"}}}},"/users/{id}":{"get":{"operationId":"UserController_getUserById","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/User"}}}},"400":{"description":"Something went wrong with the user retrieval"}}},"put":{"operationId":"UserController_updateUser","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/User"}}}},"responses":{"400":{"description":"Something went wrong with the user retrieval by id"},"401":{"description":"Unauthorized"},"404":{"description":"User not found"}},"security":[{"bearer":[]}]},"delete":{"operationId":"UserController_deleteUser","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"400":{"description":"Something went wrong with the user deletion"},"401":{"description":"Unauthorized"}},"security":[{"bearer":[]}]}},"/users/login":{"post":{"operationId":"UserController_login","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/User"}}}},"responses":{"201":{"description":"Returns a bearer token for the user"}}}}}}